add images ctrl and move graphql
