+++ app/controllers/api/images_controller.rb
@@ -1,0 +1,4 @@
+class Api::ImagesController < Api::BaseController
+  def create
+  end
+end
