+++ test/controllers/api/api_controller_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class ApiControllerTest < ActionDispatch::IntegrationTest
+  # test "the truth" do
+  #   assert true
+  # end
+end
