+++ test/controllers/api/images_controller_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class Api::ImagesControllerTest < ActionDispatch::IntegrationTest
+  # test "the truth" do
+  #   assert true
+  # end
+end
